Met à jour le calcul RDCM pour le sel d'abattage
